refactor: 클립보드 복사 로직 분리(Sharing.tsx, use-copy.ts)
